FBP-301. Extract generic action handler
